Prepare Cardano.Api.ProtocolParameters for Alonzo era. In the Alonzo era we introduce 6 new protocol parameters and remove 1 protocol parameter. In anticipation of the Alonzo era, we make some modifications to the ProtocolParameters type.
minUTxOValue becomes a Maybe (This parameter gets deprecated in Alonzo)

We introduce the following but note they are currently ignored:
- adaPerUTxoByte - Cost in ada per byte of UTxO storage (replaces
minUTxOValue)
- costmdls - The cost models for plutus scripts
- prices - The prices of execution units for plutus scripts
- maxTxExUnits - Max total script execution resource  units allowed per tx
- maxBlockExUnits - Max total script execution resource units allowed per block
- maxValSize - Max size of a Value in a tx output

We introduce an era parameter in the ProtocolParameters type in order to
reduce boilerplate. We have avoided the typical GADT pattern in
cardano-api to prevent an explosion of types.